# api/openapi.yaml
openapi: "3.0"
info:
  title: Hello World API
  version: "1.0.0"

paths:
  /sensor/ingest:
    put:
      operationId: ingestFromSensor
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestFromSensorRequest"
      responses:
        "200":
          description: Successfully ingested
  /grafana/timeseries:
    get:
      operationId: getGrafanaTimeSeries
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSeriesEntry"

components:
  schemas:
    IngestFromSensorRequest:
      type: object
      required:
        - session
        - sensorTime
        - temperature
        - humidity
      properties:
        session:
          type: string
          format: uuid
          example: f5f0e416-9144-4359-86ba-f326ff21541d
        sensorTime:
          type: string
          format: date-time
          example: 2025-06-01T00:00:00Z
        temperature:
          type: number
          format: float
          description: Temperature in celsius
          example: 20.5
        humidity:
          type: number
          format: float
          description: Relative humidity in percentage
          example: 60.5
    TimeSeriesEntry:
      type: object
      required:
        - session
        - sensorTime
        - temperature
        - humidity
      properties:
        session:
          type: string
          format: uuid
          example: f5f0e416-9144-4359-86ba-f326ff21541d
        sensorTime:
          type: string
          format: date-time
          example: 2025-06-01T00:00:00Z
        temperature:
          type: number
          format: float
          description: Temperature in celsius
          example: 20.5
        humidity:
          type: number
          format: float
          description: Relative humidity in percentage
          example: 60.5